+++ src/js/pwa/register.js
@@ -1,21 +1,23 @@
+import { showSnackbar } from '/js/ui/snackbar';
+
 export async function registerSW() {
   if (location.hostname === 'localhost' || location.hostname === '127.0.0.1' || location.hostname === '') {
-    return false; // Disallow registering service worker on localhost
+    // return false; // Disallow registering service worker on localhost
   }
 
   if ('serviceWorker' in navigator) {
     try {
-      /* navigator.serviceWorker.getRegistrations().then(function(registrations) {
+      navigator.serviceWorker.getRegistrations().then(function(registrations) {
         for (let i = 0; i < registrations.length; i++) {
           registrations[i].unregister().then(function(success) {
-            //updateSnackbar.open();
+            showSnackbar("update");
           });
         }
-      }); */
+      });
 
       navigator.serviceWorker.register('/sw.js').then(reg => {
         reg.addEventListener('updatefound', () => {
-          newWorker = reg.installing;
+          let newWorker = reg.installing;
 
           newWorker.addEventListener('statechange', () => {
             switch (newWorker.state) {